RobotTest: rearrange test code into *_test.cpp file
